pyproject.toml
[project]
name = "adzzat-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "google-adk",
    "litellm",
    "aiosmtplib",
    "python-dotenv",
]

[tool.setuptools]
packages = ["adzzat_demo"]
